Add test_host_is_bundle_loader configuration
The default behavior for a test that has a test_host is to use that test
host app as the `-bundle_loader` linker argument. This is required if
you want to test APIs from the host app, but in the case you want to use
a host for other reasons, like testing iOS features that require you to
run in an app, you might not want this. This also works around some
surprising behavior where linkopts from the app are applied to the test
bundle as well.